requirements.txt
Django==3.2.4
djangorestframework==3.12.4
channels==3.0.4
channels_redis==3.3.1
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==4.8.0
django-webpack-loader==0.7.0
